Software for Dominiks diploma thesis. In 'cdd1_local' an option with a different tolerance Lambda was added, which, however, is only active in case the macro #PROJECTOR is defined.
In 'adaptive_multiplicative_Schwarz_Proj', a projection step for 1D is added into the muliplicative
Schwarz method from Manuels thesis.

'lhs_equation' provides the matrices needed for the left-hand-side in the local problems
in the projection step (c.f. algorithm PROJECTION in Dominiks diploma thesis).

'projector_equation' similarily sets up the right-hand-side including the calculation of the
weighted integrals.

'projector_utils' provides some functions that can be used for the partition of unity needed
for the projection step.

'richardson_projector' is a Richardson-iteration in 1D including a projection step where the
local problems, too, are solved by a Richardson-Iteration

'richardson_projector_2D' the same as above, for 2D

'richardson_projector_galerkin' Richardson-Iteration with projection step. Here we use
cdd1_local as a solver for the local problems. This version was used for the numerical
experiments in Dominiks diploma thesis.

'test_multiplicative_Schwarz_Proj_1D.cpp' tests the multiplicative Schwarz method with the
projection step

'test_richardson_projector_1D' tests the Richardson-iteration using a Galerkin method for
the Poisson equation on the unit interval

'test_richardson_projector_2D' as above, but in 2D

'test_richardson_projector_constfun_2D' as above, but with constant right-hand-side